Implement DMA for eBPF device.
It can DMA both the eBPF program (the ".text" part) as well as the
".data".

To program a DMA, it is necessary to write the following (offsets from
the beginning of the MMIO region):

0x0: opcode (size: 1)
0x1: control (size: 1)
0x4: length (size: 4)
0x8: offset (size: 4)
0xc: address (size: 8)

Alternatively, the ".text" and ".data" can be used programmed by first
copying to the p2p region and then issuing a command that moves the data
from the p2p region to the appropriate regions in the device.

There are two other commands as well in this patch:
- Run program: start execution of the eBPF program
- Dump memory: dumps the content of the program and data segments
(currently, this only prints in the screen, it is not sent back to the
guest).

Signed-off-by: Martin Ichilevici de Oliveira <martin.i.oliveira@gmail.com>